backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Bryson Wu <brysonwu@student.ubc.ca>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
google-api-python-client = "^2.123.0"
python-dotenv = "^1.0.1"
motor = "^3.3.2"
newsdataapi = "^0.1.11"
beautifulsoup4 = "^4.12.3"
newspaper3k = "^0.2.8"
scrapingbee = "^2.0.1"
eventregistry = "^9.1"
python-dateutil = "^2.9.0.post0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
